Move apparmor enabled flag to a more suitable position
This moves the apparmor enabled flag to a better position under the
`podSecurity` section. I had it up in the `agent` section because it
affects things other than just the pod security policy. However that's
probably a bit more structural than these values files are supposed to
be, whereas the `podSecurity` section is more semantically linked to
what the flag does.