fix(datadog-agent): Properly set DD_APM_ENABLED in the core agent container (#1310)
The `DD_APM_ENABLED` value has always been hardcoded to `false` in the core agent
container, leading to situations where APM is indeed enabled and the value of
`apm.enabled` is set to false in the core agent and true in the trace agent. This
hasn't been an issue until we started reporting the agent configuration to the backend,
and to display it in the UI. The agent configuration is reported from the core agent,
meaning that customers will see `apm.enabled: false` in the UI even if the trace agent
exists.

Here we fix this by setting DD_APM_ENABLED to true in the core agent if the trace agent
exists or if APM is enabled. The latter should imply the former, but this makes the logic
future-proof.