APM instrumentation increase default memory limit
The default chosen in 7.57.0 of the agent is too low which results in
OOMKills for some of the APM library InitContainers. This is likely due
to a regression in the size of the APM libraries. Once a fix is merged
upstream in the Agent we should add a version gate on this setting so it
only applies to the versions affected.