feat(datadog): add lifecycle and terminationGracePeriodSeconds
This commit adds support for a couple of features which, if turned on,
apply to all agent containers generated by the chart.

1. it adds agents.lifecycle, which accepts a LifecycleHandler

ref: https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.30/#lifecyclehandler-v1-core

2. it adds agents.terminationGracePeriodSeconds, which applies to all
   agent pods if defined.

What is the motivation for this commit?

When a Kubernetes node is being shut down, it will send the TERM signal
to all non-critical pods, including Datadog. Datadog might (and often
does) finish shutting down before the application pods, which gives us
a blind spot during a very important transition.

Adding these options allows us to add, for instance, a preStop hook
which will keep the agent alive for a little longer while connections
are drained and other pods are shut down.